enhancement combining u and r options

+++ src/options/update_rules.rs
@@ -20,12 +20,12 @@ pub struct UpdateRules {}
 
 impl UpdateRules {
     /// update rules(hayabusa-rules subrepository)
-    pub fn update_rules() -> Result<String, git2::Error> {
+    pub fn update_rules(rule_path: &str) -> Result<String, git2::Error> {
         let mut result;
         let mut prev_modified_time: SystemTime = SystemTime::UNIX_EPOCH;
         let mut prev_modified_rules: HashSet<String> = HashSet::default();
         let hayabusa_repo = Repository::open(Path::new("."));
-        let hayabusa_rule_repo = Repository::open(Path::new("rules"));
+        let hayabusa_rule_repo = Repository::open(Path::new(rule_path));
         if hayabusa_repo.is_err() && hayabusa_rule_repo.is_err() {
             write_color_buffer(
                 &BufferWriter::stdout(ColorChoice::Always),
@@ -35,45 +35,57 @@ impl UpdateRules {
             )
             .ok();
             // execution git clone of hayabusa-rules repository when failed open hayabusa repository.
-            result = UpdateRules::clone_rules();
+            result = UpdateRules::clone_rules(Path::new(rule_path));
         } else if hayabusa_rule_repo.is_ok() {
             // case of exist hayabusa-rules repository
             UpdateRules::_repo_main_reset_hard(hayabusa_rule_repo.as_ref().unwrap())?;
             // case of failed fetching origin/main, git clone is not executed so network error has occurred possibly.
-            prev_modified_rules = UpdateRules::get_updated_rules("rules", &prev_modified_time);
-            prev_modified_time = fs::metadata("rules").unwrap().modified().unwrap();
+            prev_modified_rules = UpdateRules::get_updated_rules(rule_path, &prev_modified_time);
+            prev_modified_time = fs::metadata(rule_path).unwrap().modified().unwrap();
             result = UpdateRules::pull_repository(&hayabusa_rule_repo.unwrap());
         } else {
             // case of no exist hayabusa-rules repository in rules.
             // execute update because submodule information exists if hayabusa repository exists submodule information.
 
-            prev_modified_time = fs::metadata("rules").unwrap().modified().unwrap();
-            let rules_path = Path::new("rules");
+            prev_modified_time = fs::metadata(rule_path).unwrap().modified().unwrap();
+            let rules_path = Path::new(rule_path);
             if !rules_path.exists() {
                 create_dir(rules_path).ok();
             }
-            let hayabusa_repo = hayabusa_repo.unwrap();
-            let submodules = hayabusa_repo.submodules()?;
-            let mut is_success_submodule_update = true;
-            // submodule rules erase path is hard coding to avoid unintentional remove folder.
-            fs::remove_dir_all(".git/.submodule/rules").ok();
-            for mut submodule in submodules {
-                submodule.update(true, None)?;
-                let submodule_repo = submodule.open()?;
-                if let Err(e) = UpdateRules::pull_repository(&submodule_repo) {
-                    AlertMessage::alert(&format!("Failed submodule update. {}", e)).ok();
-                    is_success_submodule_update = false;
+            if rule_path == "./rules" {
+                let hayabusa_repo = hayabusa_repo.unwrap();
+                let submodules = hayabusa_repo.submodules()?;
+                let mut is_success_submodule_update = true;
+                // submodule rules erase path is hard coding to avoid unintentional remove folder.
+                fs::remove_dir_all(".git/.submodule/rules").ok();
+                for mut submodule in submodules {
+                    submodule.update(true, None)?;
+                    let submodule_repo = submodule.open()?;
+                    if let Err(e) = UpdateRules::pull_repository(&submodule_repo) {
+                        AlertMessage::alert(&format!("Failed submodule update. {}", e)).ok();
+                        is_success_submodule_update = false;
+                    }
                 }
-            }
-            if is_success_submodule_update {
-                result = Ok("Successed submodule update".to_string());
+                if is_success_submodule_update {
+                    result = Ok("Successed submodule update".to_string());
+                } else {
+                    result = Err(git2::Error::from_str(&String::default()));
+                }
             } else {
-                result = Err(git2::Error::from_str(&String::default()));
+                write_color_buffer(
+                    &BufferWriter::stdout(ColorChoice::Always),
+                    None,
+                    "Attempting to git clone the hayabusa-rules repository into the rules folder.",
+                    true,
+                )
+                .ok();
+                // execution git clone of hayabusa-rules repository when failed open hayabusa repository.
+                result = UpdateRules::clone_rules(rules_path);
             }
         }
         if result.is_ok() {
             let updated_modified_rules =
-                UpdateRules::get_updated_rules("rules", &prev_modified_time);
+                UpdateRules::get_updated_rules(rule_path, &prev_modified_time);
             result = UpdateRules::print_diff_modified_rule_dates(
                 prev_modified_rules,
                 updated_modified_rules,
@@ -129,10 +141,10 @@ impl UpdateRules {
     }
 
     /// git clone でhauyabusa-rules レポジトリをrulesフォルダにgit cloneする関数
-    fn clone_rules() -> Result<String, git2::Error> {
+    fn clone_rules(rules_path: &Path) -> Result<String, git2::Error> {
         match Repository::clone(
             "https://github.com/Yamato-Security/hayabusa-rules.git",
-            "rules",
+            rules_path,
         ) {
             Ok(_repo) => {
                 println!("Finished cloning the hayabusa-rules repository.");
